Refactor progressbar styling (#12942)
More things are shuffled around, but this is to ensure the sizing is applied
on the progressbar even if the state is negative (something is wrong with
the carrier) or if the user doesn't want to color the progressbar with signal
strenght.

While here remove a pair of useless set_size calls.

The #ifdef spaghetti for Gtk < 3.16/3.20 becomes unmaintainable.