Limit list of interfaces to IEEE 802.11 on OpenBSD and NetBSD
Contrary to interface types, media types as described in netintro(4)[0]
and ifmedia(4)[1] expose wireless interfaces as such.

Skip interfaces which do not report a IFM_IEEE80211 type via the
SIOCGIFMEDIA ioctl(2) such that the list of interfaces is reduced to
proper wifi ones alone.

FreeBSD appears to lack both an ifmedia manual and SIOCGIFMEDIA ioctl.

0: https://man.openbsd.org/netintro.4
1: https://man.openbsd.org/ifmedia.4